basic webpack library configuration
index.js and components/index.js: shims to keep the export hierarchy
neat

webpack.config.js: a basic webpack config for a library

Most of this is standard boilerplate, but there are a couple of key
things here that differ from the config for a top-level app:

+    output: {
+        ...
+        library: 'germanet-common',
+        libraryTarget: 'umd'
+    },

This tells webpack it's building a library that can be imported via
the module name 'germanet-common'. The 'umd' target tells webpack it
should build a library that is importable via the various methods
module systems: e.g., require('germanet-common') in Node.

+    externals: {
+        react: {
+            commonjs: 'react',
+            commonjs2: 'react'
+        },
+        'react-redux': {
+            commonjs: 'react-redux',
+            commonjs2: 'react-redux'
+        },
+        'react-dom': {
+            commonjs: 'react-dom',
+            commonjs2: 'react-dom'
+        },
+        redux: {
+            commonjs: 'redux',
+            commonjs2: 'redux'
+        }
+    }

This prevents webpack from bundling these dependencies into the
library.  Instead, the consuming application should install these
dependencies, and the library will look for them in that application's
node_modules directory.  This prevents, in particular, two copies of
React from being loaded (one by the library, one by the consuming
application).