helpers: remove createActionType param from makeByIdReducer
This was unnecessary; we can just assume that a defined id on an
incoming action should create a new slice of state for that id if it
isn't already there.

This change reorganizes the function a bit too, so that an error is
always thrown if a handleable action comes in without an id.