src/LexUnitDetail: draft of series of components
README: a recipe for writing a component that connect to the API
component.jsx: a bunch of components related to lexunits
actions, reducers, selectors: the usual

This work should not be considered complete.  It's a functional draft.
In particular, the code needs to be reorganized a bit, and things that
will be need in other parts of the application need to be moved to
common directories.