LexUnit components: convert to function components
Class components are more complicated, and we are not using any of the
lifecycle methods that make them necessary.  The main reason I had
defined these components as classes was to be able to define
defaultProps; but I later decided it was better to define these
defaults explicitly, especially since the only default props I was
using were limited to the case of tables.  I also eliminated the
string-based mechanism for deciding how to render lex unit data; the
user must now pass in a component as the value of the displayAs prop.
This makes the interface more consistent and makes it clearer how to
swap out a default display component for a custom one.