DataContainer: rethink and rework
This commit provides a new draft implementation of the DataContainer
component that rethinks its API.

Basically, a DataContainer is responsible for holding a set of data
and a variety of metadata about that data.  At the moment, that
metadata is *just* whether elements in the data are "chosen" or
"selected", but it may include more in the future.