DataContainer: give data selector access to the container state
This was necessary to support conditionally loading data into the
graph container based on which object was chosen.  This is a bit of a
hack around the fact that a 'data container' for a graph tree contains
only one object by design, but *which* object it should contain can
depend on the user's clicks on a previous graph.  I may want to remove
this later.