Revert "DataContainer: automatically generate ids for all containers"
This reverts commit 7b1d4e0ebb544e1da2b271dd8dab586daff35d93,
consistent with the features that have been added since then, and
adding a new warning for calling data container control methods on a
container with no .containerId.

I found, through some painful debugging, that React just can't be
relied on to keep track of these containers with automatic IDs across
renders, and this was leading to a nasty bug where control prop
actions were being emitted with the wrong ID.  The savings from having
automatically-generated IDs are really minimal, and the pain of
debugging this kind of thing is not worth it.  Thus the reversion.