D3VerticalTreeGraph: change over to nodeSize layout
This layout guarantees that the root node is at (0,0), which is what
is needed to keep the upper and lower trees in sync in
VerticalDoubleTreeGraph.

However, this layout comes with a different problem: the graph is no
longer laid out in a fixed size window, known in advance.  So we need
to recompute the viewBox and adjust accordingly.

There is still cleanup to do, but this commit represents a (nearly)
fully working example of the goal here: the graph expands in response
to clicks, the transitions work, the SVG scrolls if it goes beyond the
bounds of its containing <div>, and the viewBox gets updated properly
to keep the entire graph in view.  The next step is to make the
viewBox transition smoother and maybe add zooming.