GenericDisplay: fix unexpected behavior in TabbedPanes
The previous behavior assumed that a string was passed as tab ID, but
this is not a good assumption when TabbedPanes is used to implement
DataTabbedPanes, and data objects can have numerical IDs.  Storing the
ID as an attribute in the DOM resulted in an implicit conversion to a
string; and that in turn caused DataContainer actions like choosing to
fail, because an originally-numeric id and its stringified version are
not ===.

The fix here is to always use the original data object's ID to select
the tab, without assuming it can be stored in the DOM without
conversion.  This involves implementing the tab choosing function as a
closure, which has some minor performance implications, but I'm not
worried about them for now.