APIWrapper: do not throw error if no API call can be made at mount
component: remove the error throw in the default propsToParams
function, and wrap the call to query() in componentDidMount in a
conditional. This allows us to instantiate API-wrapped components in
situations where we don't yet have enough information to make an API
call, but we still want to render the wrapped component (for example,
to render a data container as an empty table).  This is important in
e.g. GernEdit.

The change is (more or less) backward-compatible: it essentially just
prevents an error from being thrown.  In every case where the new test
prevents a request from being made, the earlier version would have
thrown an error.

README: document the change.