APIWrapper: request on every props change
This makes the component more "functional" in the sense that any call
with a different set of props generates a new API request.

This is an important change that fixes a bug, or at least unexpected
behavior: rendering an APIWrapper-wrapped component would not cause a
new request to be generated, even when its props completely
changed. This is due a difference in the way React thinks about the
lifetime of class components, and the way that I think about
API-wrapped data containers, which are pure functions inside the
wrappers: whereas I think "new props, new container instance", React
only thinks "new props, time to re-render the existing container".  It
doesn't generate a new container instance, so it doesn't run
componentDidMount again.

Given that we *must* use the lifecycle methods, and thus class-style
components, to make API requests, the only solution to this is to make
a new request every time the props change.  This change isn't really
backwards compatible, but it provides the behavior I was expecting all
along.

The change adds a new peer dependency on lodash, for the sake of doing
prop comparisons. This is not really a new dependency; lodash is
already in the transitive dependencies of our other
dependencies (namely babel).