SynsetSearch: upgrade SynsetSearchHistoryBox to a class component
The lifecycle methods are needed to properly deal with persisting
history. In particular, we don't want to have to rely on someone else
to call reloadHistory(); so we do it in componentDidMount.  This
change also moves the storing of history into componentDidUpdate,
which provides a slight purity improvement: render now has no side
hidden effects. (The tradeoff though is that now the entire history is
persisted, and not just the filtered history.)