SynsetSearch: SynsetSearchBox -> SynsetSearchForm
SynsetSearchForm just returns a form element, which SynsetSearchBox
had previously wrapped in a Card.  This makes it more semantically
correct, as well as easier to style.  This change is NOT backward
compatible.

The change removes the alert message for a search with no results from
the form. This will become a separate component (again, for easier
styling).

Renaming the component also required renaming a bunch of things in
other places and updating the documentation.  In particular, the
various selectors for search state have changed, and the name of the
top-level reducer for searches (and the corresponding slice of the
global state) is now synsetSearches.