Graphs: implement double-click-to-reset in D3VerticalTreeGraph
The changes are relatively simple, but like everything in this
function, they took far too long to get working properly.

Mainly what I've added here is the reset() function as a handler for
double clicks.  I also abstracted the auto zooming into its own
function, so it can be called both from the main body and from the
reset handler.

I've followed this example in getting the programmatic zoom set up:
https://observablehq.com/@d3/programmatic-zoom

That led me to using svg.call(zoom.transform ...) for zoom events
instead of directly transitioning the transform attribute on the
chartContainer.  Directly transitioning the attribute was leading to
weird inconsistencies in state between programmatic and user-driven
updates to the zoom. (I'm still not sure exactly what the problem was
but things would jump around in an annoying way.  Basically, I think
the zoom behavior object didn't "know" about the programmatic updates,
and so explicit zooms would revert to an earlier transform.) The code
is now more consistent: all changes to the zoom are generated by
calling the zoom behavior's transform method, whether they are
programmatic or event-driven.

Debugging this also revealed an edge case: occasionally it's possible
for the tree to have zero width and height, leading to an infinite
scaling factor.  In practice this doesn't really matter because data
will eventually show up and trigger a redraw with a reasonable scaling
factor before anyone notices; but setting a default scaling factor of
1 in this case can still avoid errors in the browser.