Graphs: add root node ID as key in GraphSkeleton
A new root means drawing a new graph.  Without this, a graph drawn
from one root might inherit the zoom state from one drawn with a
different root.  This change has the effect of letting React know when
it needs to start fresh with a new svg element (I think).

I am not entirely sure *why* this works, but it seems to.  May need to
revisit.