ConRels: selectConRelsTree returns undefined when data unavailable
This adds a check for queryParams and queryParams.synsetId, which is
now needed because APIWrappers can be rendered even when they don't
have queryParams yet.  (The check prevents a type error when we try to
access the synset ID for the root node; this ID is needed in the
nodeFor function, and so the check has to be performed before the main
try/catch block runs.)

No other selector functions need a similar check at the moment,
because all references to queryParams are already wrapped in a
try/catch block that will return undefined if queryParams doesn't have
the required data.