GenericDisplay: TabbedPanes: add panesContainer{ClassName,Extras}
This allows extending and overriding the classes on the container
which wraps the panes divs, so you can e.g. define the width of each
tab pane as a percentage (since relative widths require a width to be
set on the parent).

This turned out to be important in GernEdit, where the tabbed panes
component that makes up the bottom row of the editor should have a
fixed width.