LexExamples: don't display ndash in list item without frame type
Not every example has an associated frame type; many do not.  It looks
funny to display the ndash, in ExampleAsListItem, when the frame type
is absent.  This change fixes that.

I had to encode the ndash as a unicode character in the string.  JSX
allows HTML entities as literals but not as strings in dynamic
content.  See https://shripadk.github.io/react/docs/jsx-gotchas.html