APIWrapper actions: rename _FETCH_ERROR -> _QUERY_ERROR
This is a tiny change; nothing is currently relying on these actions
having 'FETCH' as opposed to 'QUERY' in their type string.

I'm doing this to reserve the word 'FETCH' for API requests that
retrieve a single object, and should be distinguished from queries,
which can return zero or many objects. I anticipate adding support for
such requests soon.