DataContainer: add support for sorting
I decided to go with a fairly general mechanism here, just allowing
the user to pass a comparison function to sort the data in a (row)
container.  This seemed like the best approach because sorting e.g. by
field + an ascending/descending flag would assume that the data in the
container was structured in a certain way; and anyway, keeping track
of a sorting comparison function is simpler to implement.  To reduce
the work for the user to create sorting comparison functions in simple
cases, there is a new function, comparisonOn, in helpers.js.

Of course, it's still possible to sort the data manually inside the
display component.  What this change adds is the ability to make the
sorting *dependent on user interactions*, e.g., clicking sort buttons
or selecting sorting options from a dropdown.