GenericDisplay: add basic support for sorting in DataTables
Add a sortFields prop in DataTable, which is passed on to
DataTableHeader; this will add two buttons for ascending and
descending sort in the corresponding field's column header.  Relies on
the new .sortWith data container control prop.

This is a pretty minimal start and will probably need further
development, especially as regards styling.  But for now it meets my
needs in GernEdit.