Move away from explicitly passing data container and display props
I was originally doing this because it made the code more
self-documenting, and because it reduced the risk of unintentionally
passing on props you don't mean to pass on that could have the wrong
meaning further down in the hierarchy (e.g., className or extras).
But it's a huge pain to update all these individual files whenever I
add a feature to the underlying generic display components, and it
also makes the code less readable. So I've reversed my decision. This
change switches to using {...props} to pass props down in all
the *AsTable and *AsList components for the germanet data types.