GenericDisplay: handle arrays in DataTableRow
This change deals with the potentially-common case of displaying an
array of data in a single table cell.  It attempts to improve on the
default mostly by adding a space between the comma and the next item.

This was needed in GernEdit because ILI Records have an array in their
.synonyms field, and I can't override the displayItemAs component in
the table without breaking out of an (otherwise very nice)
abstraction.  The case seems common enough that it's worth handling
specially at this level.