GenericForms: support wrapping the main input elements in divs
Individual form elements stack vertically in Bootstrap; but without a
wrapping .form-group div, the vertical spacing between an input
element and the label of the next input element is not good, even
confusing.

This change introduces an asGroup prop to all the main input
components to allow automatically wrapping them in such a div. This
will hopefully reduce a lot of noise in individual forms.

This is done by means of the FragmentOrGroup helper component, which
is newly defined but for now not exported.

README: document the asGroup prop