GenericForms FormAlerts: errors -> error
In normal use I think validation errors are best displayed as
warnings, and error alerts should be reserved for a failure on the
server side. But this means there will usually just be one error
message, not an array of them. This change makes it easier to
construct that error message from a ManagedForm's formState: you no
longer have to wrap it in an array.