SynsetSearch: upgrade using new features from GenericForms
Use <Form> instead of <form>, and submit the form using its submitTo
prop.  The main advantage of this is that we get browser validation on
the search term field that prevents submission, instead of having to
validate that the search term is nonempty ourselves. This also meant
replacing the Button with a proper SubmitButton, and removing its
onClick handler, which was unconditionally submitting the form.

I've converted the search term input to an uncontrolled component and
used required={true} to get browser validation that prevents running a
search with an empty term. This has the benefit of not causing every
keystroke to generate a Redux action. But in order to preserve the
previous behavior of clearing the input after a normal submission and
not having the newly-empty field marked as invalid, we set the key on
the Form to props.length.history; this has the effect of recreating
the form after every search, resetting the value and the validity
state of the search term input.

I've also made use of asGroup on the Checkbox, which improves the
layout of the form slightly on smaller screens (the checkbox and its
label then wrap to the next line together, instead of the label moving
before the checkbox, when the form doesn't all fit in one row).
The checkbox remains a controlled component, because I want its state
to be preserved across searches.