Graphs: only redraw on forceRedraw *changes*
This fixes an issue introduced by commit c9c3c78: we want to force a
redraw when the forceRedraw prop is true, but only when it was false
in the immediately-previous props. Otherwise we might redraw too
often, because the Graph component can be rendered multiple times
after the prop flips from false to true. This could have been
contributing to some jerkiness in the graph rendering, though I'm not
sure that's really there.