GenericForms ManagedForm: clear most form state at submission
The form should basically reset to its default state at submission,
except that .submitting should be true, and .submittedData should get
the value of the validated data.

In other words, once the form has passed front-end validation and is
being submitted, the validation errors and server errors and response
should be cleared; otherwise, bits of state related to a previous
submission attempt can be shown alongside a success message from the
current submission.