SynsetSearch selectors: bugfix in selectSearchFormState
There was a logic error in this function that only appeared when we
started putting multiple search forms on the same page: the function
needs to return the defaultSearchFormState whenever the state for
the *given search form ID* has not yet been initialized, and not just
when there are no search forms whatsoever. The previous version was
causing a crash once one search form had associated state, because in
that case, undefined was returned as the state for a second search box.