GenericForms ManagedForm: make formState available to onSuccess
The reason we need this is that we need place to run code that runs
side effects to update the UI after the form has been successfully
submitted which has access to the form state and submitted data, since
those UI updates can depend on these.

For example, after submitting a form that deletes an object, the form
data is stale and the form should not be resubmitted. But we might not
know that the user chose to delete the object until the form was
actually submitted. Thus we need a place where we can call setState to
mark the form as disabled depending on the submitted data and on
whether or not the request was successful. onSuccess is the right
place to do that.