GenericForms Options: improve behavior in Firefox
There is apparently some subtlety to the HTML <option> element. It
can't be a self-closing tag, and despite what
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/option says,
Firefox does not seem to use the label attribute to display the label
if the element's text content is empty.  So this change moves the
label into the text content, which works on both FF and Chrome.