GenericForms ManagedForm: onSuccess can now to return a thunk
I found, as I used ManagedForm to reimplement functions in GernEdit,
that only updating the local state in the form, or even in a single
parent component, was not enough. I sometimes needed to
update *non-local* UI after a successful form submission. (Example:
adding a new hyponym to a synset should also update the con rels
table.) So I found myself wanting to dispatch Redux actions that can
induce these non-local UI updates.

This change enables that. An onSuccess handler *may* still manipulate
the form state or do whatever else it has access to; but in the
special case where it returns a function, ManagedForm will assume that
it's a Redux action creator thunk, and dispatch it.