GenericForms and Display: improve props interface of {Data}Select
DataSelect wasn't passing all the right props down to Select, and
Select was doing too much work with choose. This change moves the
construction of the onChange handler from the choose prop up to
DataSelect, since it is DataSelect that should be used as a display
component for data containers.