APIWrapper: generalize makeQueryActions
Added support for an optional dataValidator argument in
makeQueryActions. This is a function which can be used to test whether
API responses are in the expected format.

Up to now, we have always expected the data returned by the server to
be an array of objects. That expectation is now encoded in the default
dataValidator function, dataIsArrayOfObjects.

This generalization prepares the way for endpoints to return other
types of data. In particular, graph-related endpoints need to return
objects with 'nodes' and 'edges' properties, instead of arrays.

It also prepares the way for us to do stricter validation on server
responses for each of the defined germanet data types, which will
hopefully make backend bugs a bit easier to catch.

top level errors.js: add new ExternalError base class to represent
server errors (this is so that, in particular, APIError does not
extend InternalError).

APIWrapper/validation.js: define new APIError class to be thrown by
data validation functions; define default dataIsArrayOfObjects
validator.

APIWrapper/actions.js: add dataValidator as parameter of
makeQueryActions and check whether it throws an APIError before
dispatching the returned data.