SynsetSearchForm: hook up advanced options to form submission, history
Also a bit of code cleanup in SynsetSearchForm.

This required more work than I thought it would. It's easy to save the
search parameters into the history; but it's more difficult to reset
the form state to reflect the parameters of a previous search when a
history button is clicked. The problem is that, because the state of
some fields on the form depends on the state of others, a few of the
form fields have to be controlled components; but we can't control
their value via local state (as I was doing before) and at the same
time easily reset their values from a history item, which ultimately
comes in through props. The solution is to manage the state of these
form fields in Redux instead of in local state.

Thus, this change introduces a couple of new actions and corresponding
cases to handle them in the reducers:
  - UPDATE_SEARCH_PARAMS
  - TOGGLE_REGEX_SUPPORT
  - TOGGLE_CATEGORY

It also reshapes the state associated with a search form, storing all
search parameters in a single params object, which can then be
overwritten from a history item. These values are used in both
SynsetSearchForm and SynsetSearchHistoryNav.

Since there are now so many query parameters, it makes sense to have
the action which runs the search just accept an object with all the
query parameters, including the search term. A new doAdvancedSearch
thunk action creator provides this interface; the old doSearch action
creator is now implemented as a wrapper around this for backward
compatibility.