APIWrapper: more work on data validation
actions: the dataValidator function provided to makeQueryActions is
now expected to return the validated data. This is to make validation
functions here behave the same way they do in GenericForms, when data
is moving in the other direction. It also makes validation functions
more usable because it means they can be composed.

validation: rename dataIsArrayOfObjects -> isArrayofObjects and
refactor to make use of this new composability: we compose isArray and
isObject.