Graphs: two new higher-level components to draw graphs
NetworkContainer is now purely concerned with drawing the graphs and
handling the interface with vis.js. The higher-level components deal
with the representation of the data and communicating with the
backend.

PathsBetweenGraph: wrapper around NetworkContainer to draw the paths
between two given synset IDs. This component is responsible for making
the API request, fetching the data from the Redux store, and then
passing the data down to NetworkContainer to actually be drawn. It
thus functions like a data container, though for now I have not used
data containers to implement it. (There's no need at the moment, since
we have decided against supporting click-to-expand interactions in
these graphs for now.)

PathsToRootGraph: wrapper around PathsBetweenGraph that draws all
paths between a given synset and GNROOT.