Graphs: reorganize Redux state to support error messages
actions: workaround to generate an error message when we get an empty
graph object from the backend

reducers: we now track the fetching state and any error messages from
API calls alongside the graph data

selectors: change the selector to return both data and the error
message and fetching state

HnymPathsBetweenGraph: add displayUnavailable prop to support
rendering something when graph data is unavailable.  I have emulated
the behavior of DataContainers here in anticipation of eventually
wanting to replace this component with a DataContainer.