Graphs: change defaults for arrows and multiselect
New defaults draw arrows in the other direction, from hyponym to
hypernym. This is better because it does not create the expectation
that when you see arrows coming out of a node, you're seeing all of
its hyponyms.

New defaults also enable multiselect in the interactions options. This
is good because it allows selecting and dragging multiple nodes
together.