Graphs: NetworkContainer: disable physics after stabilization
By default. This makes interacting with the laid out graphs a bit
nicer on the whole, and makes it possible (though in some cases more
tedious) to manually adjust the graph to get a nice looking layout.

The new keepPhysicsOn prop can be used to keep the physics on, which
might be useful for e.g. GernEdit.