SynsetSearch: add Clear button to reset search form
actions, reducers: new action and reducer case to handle clearing
Redux-tracked search form state

SynsetSearchForm: add a ResetButton that clears the search parameters
when clicked.

Also change the submit button text from "Find" to "Search", because
this (1) is more consistent, particularly with the UI in Rover, and
does not carry the implication that anything will be found; and (2) it
makes the button a little larger, in comparison to the "Clear" button,
which is of secondary importance.