Graphs: fix keyboard interaction bug
The default vis.js keyboard interaction options grab key events from
the whole window.  This interacts badly with the ü key on a German
keyboard layout.  In GernEdit, for example, interacting with the graph
and then trying to type ü in the search form won't work: the key is
instead interpreted as a "-" and zooms the graph, and doesn't get
added to the search term.  This change fixes that behavior.  Thanks to
Ben for finding this!