Make search forms "refinable", make Clear button work with history
This required a number of changes to how the search form state is
represented in Redux and also in how (or really, *where*) we convert
the formdata into the format required by the backend.

In Redux, in order to consistently reload the form state from a saved
history item, the fields in the params object now have the same names
as the fields in the form data.

The params object is now a front-end only representation of the form
state and cannot be passed directly as a params object to axios. This
is necessitated by the data format needed by the backend, where word
categories and classes are not individual query parameters but
comma-separated lists passed to a single parameter. The conversion
from the front-end representation to the backend one now happens in (a
helper called by) doAdvancedSearch, immediately before calling out to
axios. This removes the need to worry about the backend from inside
SynsetSearchComponent, and also makes it possible for other code to
run advanced-option searches using the same flat series of boolean
flags.