SynsetSearch reducers: maintain all state but search term on submit
This is mainly for backwards compatibility with GernEdit, where the
ignoreCase box needs to remain checked after submitting a search, even
though the search term clears. I decided that it made more sense
overall if the search term is the *only* thing that clears by default,
even in the advanced form.