DataContainer: document need for idFor function more clearly
This came up because I considered skipping adding the metadata fields
when there is no .containerId, since the addMetadata functions require
the idFor function. I thought that we could be more lenient about this
when we don't need to store IDs in the container state.

But I decided against any changes, because testing made clear that
idFor is required even when the container does not have a containerId
and associated state, because IDs are needed all over the
place (commonly, to generate keys in tables or lists). So I just
documented this better.