Graphs reducer: don't erase data on error
This allows side-loading sample data (e.g. for a tutorial) into the
store so that the graph components can be used with this sample data.

Previously, the query error that results when the API request fails
would overwrite that side-loaded data, causing the graph not to
display.  I am not sure that this solution has no unintended
consequences but it's the simplest thing for now.