Draft of new Auth module
This contains the basic framework for dealing with the backend API
being behind an authentication/authorization boundary.

The strategy is:

1. Use a custom axios instance here, in germanet-common. I couldn't
get global interceptors to work right when set up from the consuming
application -- the error interceptor just wouldn't be called for
responses to API requests.

This requires the consuming application to do a little bit of setup;
that has been documented in the README.

2. The authorization state is for now just modeled as a boolean flag
called authRequired in the state. We may want something more
complicated -- e.g. storing the URL from the 401 response so the user
can visit it -- but for now the simple solution is OK. Actions,
reducers, and selectors have been provided here for this simple model.

3. API requests toggle the authRequired flag on the assumption that
unauthorized requests will return HTTP 401, and authorized requests
will return HTTP 200.