Move auth-related code up to consuming application.
I realized that the approach I was taking with the Auth functions
wasn't going to work well. The reason is that the consuming
application needs to be in control of the axios instance. If we use an
axios instance that's internal to the library, it leads to duplication
of code, because any API requests that are made *outside* of the
library will need to make use of similar interceptors. Also, the
consuming application is in a much better position to know what the
API looks like and thus how to model auth state.

Thus, this change:

 - Removes the Auth directory and code (it has been moved into a new
   auth module in the Rover repo)
 - instead puts a default axios instance in constants.js, and imports that
   where necessary
 - exports an installAxiosInstance in constants.js that allows consuming
   app to provide a custom instance which overrides the default one
 - documents installAxiosInstance in the README