Upgrade to "vis-network" package
The vis-network package seems to have migrated (back?) to this repo:
https://github.com/visjs/vis-network

That is where new work is taking place and new versions are actively
being released.  The packaged version of this repo is now apparently
vis-network, not visjs-network (which appears to be halted at version
2.24).

This change updates to the new package, adds its peer dependencies as
dependencies of germanet-common, and fixes the import in the Graphs
library. While it looks like we now have more dependencies, that isn't
really true: the vis.js folks just moved some things that were
dependencies of visjs-network to being peer-dependencies of
vis-network. Thus we now have to install them explicitly as
dependencies, but they were already being pulled in before.