GenericDisplay: ListItem: don't compute key; forward other props to li
Per the advice at:
https://reactjs.org/docs/lists-and-keys.html#extracting-components-with-keys

the key should just be given directly to ListItem, and not computed
and sent to the underlying <li> element.

This changes the props interface of ListItem. It no longer requires an
id prop (or idFor and data). This will require changes in places where
I am currently using the id prop. Those should just be converted
to keys on the ListItem.