GenericForms: add inline prop to Checkbox
I noticed that I need to make checkboxes inline fairly often, and that
I was doing it wrong (wasn't using form-check-inline, or putting it in
the wrong place, etc.). This new prop makes it trivial to get right.

Here's the Bootstrap documentation:
https://getbootstrap.com/docs/4.3/components/forms/#inline

I also noticed that, in the case of Checkbox, there is really no need
for the asGroup prop, because Checkboxes are already wrapped in a div
that also includes the label and the feedback, if any. But I am
leaving it there for consistency with the interface of the other
components.