GenericForms: ManagedForm: don't depend on react-redux
I've changed the interface to ManagedForm slightly, so that it no
longer assumes that the submitTo and onSuccess props must/can return a
thunk to be dispatched. In the case of submitTo, none of our current
uses of ManagedForm actually need the reference to dispatch; and
although various onSuccess handlers in e.g. GernEdit do, I think it
will be straightforward to rewrite them so that they do not.

The advantage of this change is that ManagedForm was the only part of
this library that depended on redux and react-redux. The change
eliminates this dependency, making the library more lightweight and
generic, and suitable even in projects where we don't need Redux.